requirements.txt
flask
python-dotenv

# Groq LLM
groq

# Image generation
gradio-client
Pillow

# Text-to-Speech
TTS
soundfile
torch

# Utilities
numpy
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.36.2
TTS==0.22.0
openai
edge-tts